Add  addresses chain - added AddressesDbContext - added AddressRepository - added GetAddressChainQuery and GetAddressChainQueryHandler - added Constants - updated entity configurations

+++ Domain/Entities/Address/AddressHierarchy.cs
@@ -2,8 +2,8 @@ namespace Domain.Entities.Address;
 
 public sealed class AddressHierarchy
 {
-    public int Id { get; set; }
-    public int ObjectId { get; set; }
-    public int ParentObjectId { get; set; }
+    public long Id { get; set; }
+    public long ObjectId { get; set; }
+    public long ParentObjectId { get; set; }
     public required string Path { get; set; }
 }
